fixed LGTM alerts and moves signalr service to proper directory
